Add files via upload
The character is the abstract class we will be using for both the player and the enemies. The player is the class that represents Link. So far I have all four of the images, one image for each direction facing, I have it drawing the correct image, and I have movement. And the Zelda runner I have has some of the things you will need in order for the player to move and also some basics like panel stuff.